+++ packages/engine/Source/Shaders/Builtin/Constants/passVector.glsl
@@ -1,9 +1,0 @@
-/**
- * The automatic GLSL constant for {@link Pass#VECTOR}
- *
- * @name czm_passVector
- * @glslConstant
- *
- * @see czm_pass
- */
-const float czm_passVector = 15.0;